optimize homeData
