+++ src/main/java/com/scrats/rent/controller/LandlordController.java
@@ -1,20 +1,30 @@
 package com.scrats.rent.controller;
 
+import com.alibaba.fastjson.JSONObject;
 import com.scrats.rent.base.service.RedisService;
+import com.scrats.rent.common.APIRequest;
+import com.scrats.rent.common.JsonResult;
 import com.scrats.rent.common.PageInfo;
+import com.scrats.rent.common.annotation.APIRequestControl;
 import com.scrats.rent.common.annotation.IgnoreSecurity;
 import com.scrats.rent.common.exception.BusinessException;
 import com.scrats.rent.common.exception.NotAuthorizedException;
 import com.scrats.rent.entity.Building;
+import com.scrats.rent.entity.Rent;
+import com.scrats.rent.entity.RoomRenter;
 import com.scrats.rent.entity.User;
 import com.scrats.rent.service.BuildingLandlordService;
 import com.scrats.rent.service.BuildingService;
+import com.scrats.rent.service.RentService;
+import com.scrats.rent.service.RoomRenterService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -34,6 +44,10 @@ public class LandlordController {
     private BuildingService buildingService;
     @Autowired
     private BuildingLandlordService buildingLandlordService;
+    @Autowired
+    private RentService rentService;
+    @Autowired
+    private RoomRenterService roomRenterService;
 
     @IgnoreSecurity
     @GetMapping("/goHome/{landlordId}")
@@ -53,5 +67,23 @@ public class LandlordController {
 
 
         return "landlord/home";
+    }
+
+    @GetMapping("/homeData/{buildingId}")
+    @ResponseBody
+    public JsonResult homeData(@APIRequestControl APIRequest apiRequest, @PathVariable Integer buildingId){
+        if(!apiRequest.isLandlordFlag()){
+            throw new BusinessException("数据错误!");
+        }
+        Building building = buildingService.selectByPrimaryKey(buildingId);
+        List<RoomRenter> roomRenterList = roomRenterService.getRoomRenterByBuildingId(buildingId);
+        List<Rent> rentList = rentService.getRentByBuildingIdandPayFlag(buildingId, false);
+
+        JSONObject jsonObject = new JSONObject();
+        jsonObject.put("roomNum", building.getRooms());
+        jsonObject.put("avaliableNum", building.getRoomAble());
+        jsonObject.put("renterNum", roomRenterList.size());
+        jsonObject.put("noRentNum", rentList.size());
+        return new JsonResult(jsonObject);
     }
 }
